This datasets contains a set of simulations of a known benchmark in the fault diagnosis for chemical processes community, i.e., the Continuous Stirred Tank Reactor. This system carries an exothermic reaction A -> B, and a feedback loop for controlling the reactor's temperature. There are 7 measured variables,
Concentration of A in the inlet flow,
Temperature of the inlet flow,
Temperature of the inlet coolant flow,
Coolant flow-rate,
Concentration of B in the outlet flow,
Temperature of the outlet flow,
Temperature of the outlet coolant flow.
The goal is to predict a set of 12 faults from these 7 variables, measured throughout 200 minutes, at a 1 minute rate.